[change] shparos-> spharos
