--- src/main/java/shparos/review_bookmark/bookmark/vo/request/RegisterBookmarkRequest.java
@@ -1,10 +1,0 @@
-package shparos.review_bookmark.bookmark.vo.request;
-
-import lombok.Getter;
-
-@Getter
-public class RegisterBookmarkRequest {
-
-    private Long serviceId;
-
-}
