--- src/main/java/shparos/review_bookmark/bookmark/vo/response/BookmarkResponse.java
@@ -1,17 +1,0 @@
-package shparos.review_bookmark.bookmark.vo.response;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-public class BookmarkResponse {
-
-    private Long bookmarkId;
-    private Long serviceId;
-
-}
