--- src/main/java/shparos/review_bookmark/global/exception/GlobalExceptionController.java
@@ -1,19 +1,0 @@
-package shparos.review_bookmark.global.exception;
-
-import lombok.extern.slf4j.Slf4j;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.RestControllerAdvice;
-import shparos.review_bookmark.global.common.response.ErrorResponse;
-
-@Slf4j
-@RestControllerAdvice
-public class GlobalExceptionController {
-
-    @ExceptionHandler(value = { CustomException.class })
-    public ResponseEntity<?> customExHandle(CustomException e) {
-        log.error("[exceptionHandle] ex", e);
-        return ErrorResponse.toResponseEntity(e.getResponseCode());
-    }
-
-}
