+++ src/main/java/spharos/review_bookmark/bookmark/vo/request/RegisterBookmarkRequest.java
@@ -1,0 +1,10 @@
+package spharos.review_bookmark.bookmark.vo.request;
+
+import lombok.Getter;
+
+@Getter
+public class RegisterBookmarkRequest {
+
+    private Long serviceId;
+
+}
